// src/app/demo/components/pages/_core/models/convocatoria.model.ts
export interface Convocatoria {
    id:           number;
    url:          string;
    link:         string;
    imagen:       string;
    titulo:       string;
    bajada:       string;
    cuerpo:       string;
    estado:       string;
    fecha_inicio: string;
    fecha_fin:    string;
    documentos:   string[];
}